Extract subtitle frame source
